# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/promotions": {
      "get": {
        "operationId": "PromotionsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List promotions",
        "tags": [
          "Promotions"
        ]
      },
      "post": {
        "operationId": "PromotionsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an HFSS-aware promotion",
        "tags": [
          "Promotions"
        ]
      }
    },
    "/api/v1/promotions/{id}": {
      "patch": {
        "operationId": "PromotionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update promotion schedule or status",
        "tags": [
          "Promotions"
        ]
      }
    },
...
    "/api/v1/privacy/consents": {
      "get": {
        "operationId": "PrivacyController_consents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      },
      "patch": {
        "operationId": "PrivacyController_updateConsents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    },
    "/api/v1/privacy/requests": {
      "get": {
        "operationId": "PrivacyController_requests",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    },
    "/api/v1/privacy/export-requests": {
      "post": {
        "operationId": "PrivacyController_exportRequest",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    },
    "/api/v1/privacy/export-requests/{id}": {
      "get": {
        "operationId": "PrivacyController_exportFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    },
    "/api/v1/privacy/deletion-requests": {
      "post": {
        "operationId": "PrivacyController_erasureRequest",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    },
    "/api/v1/admin/privacy/requests": {
      "get": {
        "operationId": "PrivacyController_adminRequests",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    },
    "/api/v1/admin/privacy/requests/{id}": {
      "patch": {
        "operationId": "PrivacyController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacy"
        ]
      }
    }
  },
  "info": {
    "title": "Denes Commerce API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}